requirements.txt
pymysql~=1.0.2
pyyaml~=6.0
tushare~=1.2.89